Move test for successful metadata extraction into `TestExtractDsMeta`
