+++ datalad_registry/tests/test_tasks/__init__.py
@@ -1,1 +1,3 @@
 TEST_MIN_REPO_URL = "https://github.com/datalad/testrepo--minimalds.git"
+TEST_MIN_REPO_TAG = "0.1.0"
+TEST_MIN_REPO_COMMIT_HEXSHA = "ac9ba85cf1e8a004e7c24ebb6b5cd861d53e3998"
