+++ datalad_registry/tests/test_tasks/test_extract_meta.py
@@ -2,8 +2,7 @@ from datalad import api as dl
 from datalad.distribution.dataset import require_dataset
 import pytest
 
-from datalad_registry.blueprints.api.url_metadata import URLMetadataModel
-from datalad_registry.tasks import ExtractMetaStatus, extract_ds_meta, extract_meta
+from datalad_registry.tasks import ExtractMetaStatus, extract_meta
 
 _TEST_REPO_URL = "https://github.com/datalad/testrepo--minimalds.git"
 _TEST_REPO_TAG = "0.1.0"
@@ -35,43 +34,6 @@ def test_repo_path(tmp_path_factory):
 
 
 class TestExtractMeta:
-    def test_succeeded(self, flask_app, processed_ds_urls):
-        """
-        Test that metadata extraction returns ExtractMetaStatus.SUCCEEDED when
-        all provided arguments are valid, and the given extractor doesn't require
-        any special file to be present in the dataset
-        """
-        from datalad_registry.models import URL, db
-
-        test_repo_url_id = processed_ds_urls[0]
-
-        with flask_app.app_context():
-            ret = extract_ds_meta(test_repo_url_id, _BASIC_EXTRACTOR)
-
-            assert ret == ExtractMetaStatus.SUCCEEDED
-
-            url = db.session.execute(
-                db.select(URL).where(URL.id == test_repo_url_id)
-            ).scalar_one()
-
-            metadata_lst = url.metadata_
-
-            # Verify the number of pieces of metadata
-            assert len(metadata_lst) == 1
-
-            metadata = URLMetadataModel.from_orm(metadata_lst[0])
-
-            # Verify metadata saved to database
-            assert metadata.dataset_describe == _TEST_REPO_TAG
-            assert metadata.dataset_version == _TEST_REPO_COMMIT_HEXSHA
-            assert metadata.extractor_name == _BASIC_EXTRACTOR
-            assert metadata.extraction_parameter == {}
-            # noinspection HttpUrlsUsage
-            assert metadata.extracted_metadata["@context"] == {
-                "@vocab": "http://schema.org/",
-                "datalad": "http://dx.datalad.org/",
-            }
-
     def test_aborted(self, flask_app, test_repo_url_id, test_repo_path):
         """
         Test that metadata extraction returns ExtractMetaStatus.ABORTED when
